Changed general cmd file name, removed duplicate code, added better comments.
