--- cmd/api-endpoint/expected_output_test.go
@@ -1,47 +1,0 @@
-package main
-
-const SUBMISSION_1697406272 = `{
-    "found-submission": true,
-    "found-user": true,
-    "submission-result": {
-        "additional-info": null,
-        "assignment-id": "hw0",
-        "course-id": "course101",
-        "grading_end_time": 1697406273000,
-        "grading_start_time": 1697406273000,
-        "id": "course101::hw0::course-student@test.edulinq.org::1697406272",
-        "max_points": 2,
-        "message": "",
-        "name": "HW0",
-        "questions": [
-            {
-                "grading_end_time": 1697406273000,
-                "grading_start_time": 1697406273000,
-                "max_points": 1,
-                "message": "",
-                "name": "Q1",
-                "score": 1
-            },
-            {
-                "grading_end_time": 1697406273000,
-                "grading_start_time": 1697406273000,
-                "max_points": 1,
-                "message": "",
-                "name": "Q2",
-                "score": 1
-            },
-            {
-                "grading_end_time": 1697406273000,
-                "grading_start_time": 1697406273000,
-                "max_points": 0,
-                "message": "Style is clean!",
-                "name": "Style",
-                "score": 0
-            }
-        ],
-        "score": 2,
-        "short-id": "1697406272",
-        "user": "course-student@test.edulinq.org"
-    }
-}
-`
