+++ cmd/call-api-endpoint/expected_output_test.go
@@ -1,0 +1,47 @@
+package main
+
+const SUBMISSION_1697406256 = `{
+    "found-submission": true,
+    "found-user": true,
+    "submission-result": {
+        "additional-info": null,
+        "assignment-id": "hw0",
+        "course-id": "course101",
+        "grading_end_time": 1697406256000,
+        "grading_start_time": 1697406256000,
+        "id": "course101::hw0::course-student@test.edulinq.org::1697406256",
+        "max_points": 2,
+        "message": "",
+        "name": "HW0",
+        "questions": [
+            {
+                "grading_end_time": 1697406256000,
+                "grading_start_time": 1697406256000,
+                "max_points": 1,
+                "message": "NotImplemented returned.",
+                "name": "Q1",
+                "score": 0
+            },
+            {
+                "grading_end_time": 1697406256000,
+                "grading_start_time": 1697406256000,
+                "max_points": 1,
+                "message": "NotImplemented returned.",
+                "name": "Q2",
+                "score": 0
+            },
+            {
+                "grading_end_time": 1697406256000,
+                "grading_start_time": 1697406256000,
+                "max_points": 0,
+                "message": "Style is clean!",
+                "name": "Style",
+                "score": 0
+            }
+        ],
+        "score": 0,
+        "short-id": "1697406256",
+        "user": "course-student@test.edulinq.org"
+    }
+}
+`
