+++ cmd/course-user-list/main.go
@@ -1,6 +1,8 @@
 package main
 
 import (
+	"strings"
+
 	"github.com/alecthomas/kong"
 
 	"github.com/edulinq/autograder/internal/api/core"
@@ -8,6 +10,7 @@ import (
 	"github.com/edulinq/autograder/internal/cmd"
 	"github.com/edulinq/autograder/internal/config"
 	"github.com/edulinq/autograder/internal/log"
+	"github.com/edulinq/autograder/internal/util"
 )
 
 var args struct {
@@ -36,8 +39,34 @@ func main() {
 
 	var printFunc cmd.CustomResponseFormatter = nil
 	if args.Table {
-		printFunc = cmd.ListCourseUsersTable
+		printFunc = listCourseUsersTable
 	}
 
 	cmd.MustHandleCMDRequestAndExitFull(`courses/users/list`, request, users.ListResponse{}, args.CommonOptions, printFunc)
+}
+
+func listCourseUsersTable(response core.APIResponse) string {
+	var responseContent users.ListResponse
+	util.MustJSONFromString(util.MustToJSON(response.Content), &responseContent)
+
+	var courseUsersTable strings.Builder
+
+	headers := []string{"email", "name", "role", "lms-id"}
+	courseUsersTable.WriteString(strings.Join(headers, "\t") + "\n")
+
+	for i, user := range responseContent.Users {
+		if i > 0 {
+			courseUsersTable.WriteString("\n")
+		}
+
+		courseUsersTable.WriteString(user.Email)
+		courseUsersTable.WriteString("\t")
+		courseUsersTable.WriteString(user.Name)
+		courseUsersTable.WriteString("\t")
+		courseUsersTable.WriteString(user.Role.String())
+		courseUsersTable.WriteString("\t")
+		courseUsersTable.WriteString(user.LMSID)
+	}
+
+	return courseUsersTable.String()
 }
